Refactor: Match namespace and path as it should and better namespace
