+++ tests/_support/Steps/WebsiteUserSteps.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace GigaScience;
+namespace Steps;
 
 /**
  * Class WebsiteUserSteps
